Sync Health Center agent src Agent 	78438: Node: Cannot enable logging on Node plugins - Make logging configuration work in Node plugins 	78432: fix logic of nodynamic trace 	78437: Node: CPU plugin pull interval too frequent - Change CPU pull interval from 1s to 2s 	78294: OOM on Linux PPC32/64 in 3hr headless 	78175: Node: Check and fix types - More type fixes, particularly monitordata.data which is a uint32 rather than size_t. 	78106: Node: Code cleanup - Remove unused includes 	78175: Node: Check and fix types - Replace sprintf with stringstream. Fix includes. Fix types and casts. Other small fixes. 	78106: Node: Code cleanup - Use itoa() from common.h where appropriate. Prepend std:: namespace where appropriate. 	78106: Node: Code cleanup - Remove zos specific code from within aix/windows specific sections 	78106: Node: Code cleanup - Use C++ casts in cpu plugin 	78290: Odd characters on zOS from agent start 	78130: Strange headless files in performance runs - honour user buffer size 	78130: Strange headless files in performance runs 	76602: Node: Fix compiler warnings - Fix types and remove sprintf/sscanf/fopen/fread in cpu plugin 	78130: Strange headless files in performance runs - Force expire the counters or immediate update only gets the first 5 	78130: Strange headless files in performance runs

+++ runtime.tools/src/ibmras/monitoring/plugins/cpu/cputime.h
@@ -11,11 +11,16 @@
 #ifndef ibmras_monitoring_plugins_cpu_cputime_h
 #define ibmras_monitoring_plugins_cpu_cputime_h
 
+#include "ibmras/common/types.h"
+
+// TODO Should rewrite this as a class and not expose timestamp
+//      use the platform specific time values in an opaque way
+//      as converting them to uint64s is problematic
 struct CPUTime {
-	unsigned long long time; /* ns since fixed point */
-	unsigned long long total; /* cumulative total cpu time in ns */
-	unsigned long long process; /* cumulative process cpu time in ns */
-	int nprocs;
+	uint64 time; /* ns since fixed point */
+	uint64 total; /* cumulative total cpu time in ns */
+	uint64 process; /* cumulative process cpu time in ns */
+	uint32 nprocs;
 };
 
 extern "C" struct CPUTime* getCPUTime();
